[BE] ADD : 댓글 조회 추가

+++ backend/src/main/java/com/ssafy/jupging/dto/CommentSaveRequestDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CommentSaveRequestDto {
-    private int articleId;
-    private int userId;
+    private Long articleId;
+    private Long userId;
     private String commentContent;
 }
